Added JUnit test
