+++ src/test/java/seedu/planitarium/money/IncomeTest.java
@@ -1,50 +1,58 @@
-package seedu.planitarium.expenditure;
+//@@author tjiarong
+
+package seedu.planitarium.money;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import seedu.planitarium.money.Expenditure;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.fail;
 
-class ExpenditureTest {
+class IncomeTest {
 
-    private Expenditure testItem;
+    private Income testItem;
 
     @BeforeEach
     public void setUp() {
-        testItem = new Expenditure("food", 24, 1,false);
+        testItem = new Income("food", 24, false);
     }
 
     @Test
-    public void getDescription_validExpenditure_expectedDescription() {
+    public void getDescription_validIncome_expectedDescription() {
         String inputDescription = "food";
         String actualDescription = testItem.getDescription();
         assertEquals(inputDescription, actualDescription);
     }
 
     @Test
-    public void getAmount_validExpenditure_expectedAmount() {
+    public void getAmount_validIncome_expectedAmount() {
         double inputAmount = 24;
         double actualAmount = testItem.getAmount();
         assertEquals(inputAmount, actualAmount);
     }
 
     @Test
-    public void getPermanent_validExpenditure_expectedPermanent() {
+    public void getPermanent_validIncome_expectedPermanent() {
         boolean isPermanent = testItem.isPermanent();
         assertFalse(isPermanent);
     }
 
     @Test
-    public void addExpenditure_nullDescription_expectAssertionError() {
+    public void addIncome_nullDescription_expectAssertionError() {
         try {
-            Expenditure testNullDescription = new Expenditure(null, 24, 1,false);
+            Income testNullDescription = new Income(null, 24, false);
             fail();
         } catch (AssertionError e) {
             assertNull(e.getMessage());
         }
+    }
+
+    @Test
+    public void saveString_getString_success() {
+        String result = "i food - 24.0 - false";
+        String actual = testItem.saveString();
+        assertEquals(result, actual);
     }
 }
